requirements.txt
tweepy
pyyaml
pymysql
requests
boto3
fastai
torch==1.0.0
torchvision==0.2.2